requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
SQLAlchemy==2.0.25
python-dotenv==1.0.1
pydantic==1.10.13
python-dateutil==2.8.2
alembic==1.12.1
pytest==7.4.4
freezegun==1.4.0
humanize==4.9.0
